backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
pandas==1.3.5
requests==2.31.0
gunicorn==20.1.0
python-dotenv==0.21.1